pyproject.toml
[project]
name = "edm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=5.4.0",
    "matplotlib>=3.10.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
